search_service/pyproject.toml
[project]
name = "search-service"
version = "0.1.0"
description = "ASTU Connect Search Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-jose[cryptography]",
    "pydantic-settings",
    "aiokafka",
    "elasticsearch>=8,<9",
]
